+++ Poseidon/Poseidon/Enemies/ShootingEnemy.cs
@@ -26,7 +26,7 @@ namespace Poseidon
             damage = GameConstants.EnemyShootingDamage;
             if (PoseidonGame.gamePlus)
             {
-                timeBetweenFire /= (1 + HydroBot.gamePlusLevel * 0.25f);
+                timeBetweenFire /= (1 + (float)HydroBot.gamePlusLevel * 0.25f);
                 damage *= (HydroBot.gamePlusLevel + 1);
             }
